Add files via upload
*What's New*:

Quite a lot, where to begin?

- Animations implemented to protagonist, mobs and farmer (lots of drawing involved)
- Multi-layer backgrounds with varying scrolling sensitivities
- Level 3 completely redone with new stage
- Default plot timeline extended with additional level (similar as the course followed by the original level 3 of this timeline, but better)
- Speechboxes and icons added for dialogue
- Customisable username, which is implemented in the dialogue
- Fixed bug of player disappearing below canvas
- Fixed bug of player "hitting a ceiling" as the vertical scrolling is toggled on 

Due to the time and complexity it took me to draw the many image files and tweak the coding, I was only able to complete a single plot timeline so far. As a result, to prevent the game from crashing, I had to disable the alternative plots for now (in other words, only one ending, the sky/sea, beach and cyber/nostalgia levels are inaccessible in this version). Hopefully, I will have them all ready next month, although I cannot promise anything.